fix: correct bandpass filter coefficients and match rsudp filtering behavior
Fix Butterworth bandpass SOS computation that produced 8 sections instead
of 4 by eliminating conjugate pole duplication. Apply forward-only filter
to full deconvolved context buffer matching rsudp's obspy default
(zerophase=False). Add instrument response deconvolution via FFT spectral
division with dynamic context sizing.

+++ webui/lib/types.ts
@@ -49,7 +49,7 @@ export interface AlertSettings {
 }
 
 export interface SpectrogramColumn {
-  data: Uint8Array;
+  data: Float32Array;
   timestamp: number; // microseconds since epoch
   frequencyBins: number;
 }
@@ -60,7 +60,7 @@ export interface SpectrogramPacket {
   sampleRate: number;
   frequencyBins: number;
   columnsCount: number;
-  columns: Uint8Array[];
+  columns: Float32Array[];
 }
 
 export interface IntensityIndicatorState {
